requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
bleach==1.5.0
chardet==2.3.0
coreapi==2.2.4
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.5
django-allauth==0.30.0
django-contrib-comments==1.8.0
django-cors-headers==2.0.2
django-filter==1.0.1
django-material==0.12.5
django-postgrespool==0.3.1
django-rest-auth==0.9.0
django-rest-swagger==2.1.1
django-taggit==0.22.0
django-taggit-serializer==0.1.5
django-versatileimagefield==1.6.3
djangorestframework==3.5.4
drfdocs==0.0.11
Drum==0.4.0
filebrowser-safe==0.4.6
future==0.16.0
grappelli-safe==0.4.5
gunicorn==19.6.0
html5lib==0.9999999
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
Mezzanine==4.2.2
oauthlib==2.0.1
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python3-openid==3.0.10
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
sphinx-me==0.3
SQLAlchemy==1.1.5
tzlocal==1.3
uritemplate==3.0.0
whitenoise==3.3.0
